Removed login and post features, for now.
- Using new way of creating the header.

+++ src/servrian/post.h
@@ -15,7 +15,7 @@
 #include "aux.h"
 
 /* This function responds to a POST request */
-int handle_post(unsigned int conn, struct response res);
+short handle_post(short conn, struct response res);
 
 #endif
